% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{svyflow_results}
\alias{svyflow_results}
\title{The \code{svyflow_results} class}
\description{
\code{\link[=analyze_survey]{analyze_survey()}} returns a tibble of class \code{svyflow_results}. The class
is a hook for future output formatters (Excel, HTML, publication-style
wide tables) which will dispatch on it via S3. There is no functional
difference from a regular tibble today besides the \code{\link[=print]{print()}} header.
}
\keyword{internal}
